Create main.py
main.py calls helper functions. 
James Implementation: determines the INNER GOLD ellipse and TWO candidates.

if no ellipse, ellipes is set to None, and candidates are zeroes. 
Otherwise the two candidate poses are in mm, and are in camera coordinates. IE the distances represent distances from the camera optical center to the circle plane.